requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.6.15
chardet==5.2.0
charset-normalizer==3.4.2
Django==5.2.3
django-cors-headers==4.7.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
idna==3.10
mysqlclient==2.2.7
pillow==11.2.1
promise==2.3
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.9.0
PyMySQL==1.1.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
reportlab==4.4.1
requests==2.32.4
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
wheel==0.45.1
